Implementing 2d (row x elen) multiple table adapter for assembly reads
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@367 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/include/U2Core/U2Attribute.h
@@ -46,7 +46,7 @@ public:
     U2DataId    childId;
 
     /** parent object version this attribute is related to. If <=0 -> any is OK. */
-    qint32      version;
+    qint64      version;
     
     /* Name of the attribute */
     QString     name;
